codecon/requirements.txt
adium-theme-ubuntu==0.3.4
appdirs==1.4.0
astroid==1.4.9
backports.functools-lru-cache==1.3
beautifulsoup4==4.5.3
chardet==2.3.0
configparser==3.5.0
coreapi==2.3.0
coreschema==0.0.4
Django==1.10.5
django-allauth==0.31.0
django-modelcluster==2.0
django-rest-auth==0.9.1
django-rest-swagger==2.0.7
django-taggit==0.18.3
django-treebeard==4.1.0
djangorestframework==3.5.3
gyp==0.1
html5lib==0.999999999
httpie==0.9.9
isort==4.2.5
itypes==1.1.0
Jinja2==2.9.6
lazy-object-proxy==1.2.2
lxml==3.5.0
MarkupSafe==1.0
mccabe==0.6.1
oauthlib==2.0.2
olefile==0.44
openapi-codec==1.3.1
packaging==16.8
Pillow==4.0.0
psycopg2==2.7.1
Pygments==2.2.0
pygobject==3.20.0
pylint==1.6.5
pyparsing==2.1.10
python-openid==2.2.5
pytz==2016.10
requests==2.13.0
requests-oauthlib==0.8.0
simplejson==3.10.0
six==1.10.0
Unidecode==0.4.20
unity-lens-photos==1.0
uritemplate==3.0.0
wagtail==1.8.1
webencodings==0.5
Willow==0.4
wrapt==1.10.10
zenmap==7.1
